+++ .idea/modules.xml
@@ -2,7 +2,11 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+<<<<<<< HEAD
       <module fileurl="file://$PROJECT_DIR$/LLapp-master.iml" filepath="$PROJECT_DIR$/LLapp-master.iml" />
+=======
+      <module fileurl="file://$PROJECT_DIR$/GitHub.iml" filepath="$PROJECT_DIR$/GitHub.iml" />
+>>>>>>> c0983f57036b4d4d25f33530bead12cd85107b11
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
