+++ LLapp-master/local.properties
@@ -1,0 +1,11 @@
+## This file is automatically generated by Android Studio.
+# Do not modify this file -- YOUR CHANGES WILL BE ERASED!
+#
+# This file must *NOT* be checked into Version Control Systems,
+# as it contains information specific to your local configuration.
+#
+# Location of the SDK. This is only used by Gradle.
+# For customization when using a Version Control System, please read the
+# header note.
+#Sat Mar 25 15:05:30 GMT 2017
+sdk.dir=C\:\\Users\\Dragos\\AppData\\Local\\Android\\Sdk
