+++ app/src/main/java/georgia/languagelandscape/SignUpActivity.java
@@ -124,7 +124,7 @@ public class SignUpActivity extends AppCompatActivity {
 
     public void sendMessage(JSONObject object){
         try{
-            String url = "http://a50f184b.ngrok.io";
+            String url = "http://7602f906.ngrok.io/";
             JsonObjectRequest req = new JsonObjectRequest(url, object,
                     new Response.Listener<JSONObject>() {
                         @Override
