+++ app/src/main/res/layout/content_navigation_drawer.xml
@@ -5,10 +5,6 @@
     android:id="@+id/content_navigation_drawer"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:paddingBottom="@dimen/activity_vertical_margin"
-    android:paddingLeft="@dimen/activity_horizontal_margin"
-    android:paddingRight="@dimen/activity_horizontal_margin"
-    android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
     tools:context="georgia.languagelandscape.BaseActivity"
     tools:showIn="@layout/app_bar_navigation_drawer">
